@@ -8,31 +8,66 @@
     "eject": "expo eject"
   },
   "dependencies": {
-    "@react-native-community/masked-view": "^0.1.10",
+    "@ffprobe-installer/ffprobe": "^1.1.0",
+    "@react-native-community/cameraroll": "^4.0.1",
+    "@react-native-community/masked-view": "0.1.10",
+    "@react-native-community/netinfo": "6.0.0",
+    "@react-native-community/slider": "3.0.3",
     "@react-navigation/bottom-tabs": "^5.8.0",
     "@react-navigation/material-bottom-tabs": "^5.2.16",
+    "@react-navigation/material-top-tabs": "^5.3.13",
     "@react-navigation/native": "^5.7.3",
     "@react-navigation/stack": "^5.9.0",
-    "expo": "~38.0.8",
-    "expo-status-bar": "^1.0.2",
+    "@skele/components": "^1.0.0-alpha.40",
+    "babel-preset-expo": "^8.3.0",
+    "expo": "~42.0.3",
+    "expo-av": "~9.2.3",
+    "expo-camera": "~11.2.2",
+    "expo-cameraroll": "^1.1.0",
+    "expo-checkbox": "^1.0.3",
+    "expo-device": "~3.3.0",
+    "expo-image-picker": "~10.2.2",
+    "expo-linear-gradient": "~9.2.0",
+    "expo-media-library": "~12.1.2",
+    "expo-notifications": "~0.12.3",
+    "expo-status-bar": "~1.0.4",
+    "expo-updates": "~0.8.3",
+    "expo-video-player": "^2.0.1",
+    "expo-video-thumbnails": "~5.2.1",
     "firebase": "8.2.3",
-    "react": "~16.11.0",
-    "react-dom": "~16.11.0",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-38.0.2.tar.gz",
-    "react-native-gesture-handler": "^1.8.0",
-    "react-native-paper": "^4.2.0",
-    "react-native-screens": "^2.10.1",
+    "get-video-duration": "^3.1.0",
+    "queue": "^6.0.2",
+    "react": "16.13.1",
+    "react-dom": "16.13.1",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-42.0.0.tar.gz",
+    "react-native-bottomsheet-reanimated": "0.0.22",
+    "react-native-elements": "^3.0.0-alpha.1",
+    "react-native-expo-cached-image": "^1.3.1",
+    "react-native-fast-image": "^8.3.4",
+    "react-native-gesture-handler": "~1.10.2",
+    "react-native-interactable-reanimated": "0.0.8",
+    "react-native-mentions": "^1.1.4",
+    "react-native-optimized-flatlist": "^1.0.4",
+    "react-native-paper": "^4.7.1",
+    "react-native-parsed-text": "0.0.22",
+    "react-native-reanimated": "~2.2.0",
+    "react-native-restart": "0.0.20",
+    "react-native-screens": "~3.4.0",
+    "react-native-tab-view": "^2.15.2",
+    "react-native-unimodules": " ~0.14.6",
     "react-native-vector-icons": "^7.1.0",
-    "react-native-web": "~0.11.7",
+    "react-native-web": "~0.13.12",
     "react-redux": "^7.2.1",
+    "react-uuid": "^1.0.2",
+    "react-visibility-sensor": "^5.1.1",
+    "reanimated-bottom-sheet": "^1.0.0-alpha.22",
     "redux": "^4.0.5",
     "redux-thunk": "^2.3.0",
-    "expo-camera": "~8.3.1",
-    "expo-image-picker": "~8.3.0"
+    "sentry-expo": "^4.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.8.6",
-    "babel-preset-expo": "~8.1.0"
+    "@babel/core": "~7.13.0",
+    "babel-plugin-module-resolver": "^4.1.0"
   },
   "private": true
 }
