@@ -1,38 +1,150 @@
-import { USER_STATE_CHANGE, USER_POSTS_STATE_CHANGE, USER_FOLLOWING_STATE_CHANGE, USERS_DATA_STATE_CHANGE,USERS_POSTS_STATE_CHANGE, USERS_LIKES_STATE_CHANGE, CLEAR_DATA} from '../constants/index'
-import firebase from 'firebase'
-import { SnapshotViewIOSComponent } from 'react-native'
+import * as Notifications from 'expo-notifications';
+import firebase from 'firebase';
+import { Constants } from 'react-native-unimodules';
+import { CLEAR_DATA, USERS_DATA_STATE_CHANGE, USERS_LIKES_STATE_CHANGE, USERS_POSTS_STATE_CHANGE, USER_CHATS_STATE_CHANGE, USER_FOLLOWING_STATE_CHANGE, USER_POSTS_STATE_CHANGE, USER_STATE_CHANGE } from '../constants/index';
 require('firebase/firestore')
 
 
+let unsubscribe = [];
+
 export function clearData() {
     return ((dispatch) => {
-        dispatch({type: CLEAR_DATA})
+        for (let i = unsubscribe; i < unsubscribe.length; i++) {
+            unsubscribe[i]();
+        }
+        dispatch({ type: CLEAR_DATA })
     })
 }
-export function fetchUser() {
+export function reload() {
     return ((dispatch) => {
-        firebase.firestore()
-            .collection("users")
-            .doc(firebase.auth().currentUser.uid)
-            .get()
-            .then((snapshot) => {
-                if (snapshot.exists) {
-                    dispatch({ type: USER_STATE_CHANGE, currentUser: snapshot.data() })
-                }
-                else {
-                    console.log('does not exist')
-                }
-            })
+        dispatch(clearData())
+        dispatch(fetchUser())
+        dispatch(setNotificationService())
+        dispatch(fetchUserPosts())
+        dispatch(fetchUserFollowing())
+        dispatch(fetchUserChats())
+
     })
 }
 
+export const setNotificationService = () => async dispatch => {
+    let token;
+    if (Constants.isDevice) {
+        const existingStatus = await Notifications.getPermissionsAsync();
+        let finalStatus = existingStatus;
+        if (existingStatus.status !== 'granted') {
+            const status = await Notifications.requestPermissionsAsync();
+            finalStatus = status;
+        }
+
+        if (finalStatus.status !== 'granted') {
+            alert('Failed to get push token for push notification!');
+            return;
+        }
+        token = (await Notifications.getExpoPushTokenAsync());
+    } else {
+        alert('Must use physical device for Push Notifications');
+    }
+
+    if (Platform.OS === 'android') {
+        Notifications.setNotificationChannelAsync('default', {
+            name: 'default',
+            importance: Notifications.AndroidImportance.MAX,
+            vibrationPattern: [0, 250, 250, 250],
+            lightColor: '#FF231F7C',
+        });
+    }
+
+    Notifications.setNotificationHandler({
+        handleNotification: async () => ({
+            shouldShowAlert: true,
+            shouldPlaySound: false,
+            shouldSetBadge: false,
+        }),
+    });
+
+    if (token != undefined) {
+        firebase.firestore()
+            .collection("users")
+            .doc(firebase.auth().currentUser.uid)
+            .update({
+                notificationToken: token.data,
+            })
+    }
+
+}
+
+export const sendNotification = (to, title, body, data) => dispatch => {
+    if (to == null) {
+        return;
+    }
+
+    let response = fetch('https://exp.host/--/api/v2/push/send', {
+        method: 'POST',
+        headers: {
+            Accept: 'application/json',
+            'Content-Type': 'application/json'
+        },
+        body: JSON.stringify({
+            to,
+            sound: 'default',
+            title,
+            body,
+            data
+        })
+    })
+
+}
+
+export function fetchUser() {
+    return ((dispatch) => {
+        let listener = firebase.firestore()
+            .collection("users")
+            .doc(firebase.auth().currentUser.uid)
+            .onSnapshot((snapshot, error) => {
+                if (snapshot.exists) {
+                    dispatch({ type: USER_STATE_CHANGE, currentUser: { uid: firebase.auth().currentUser.uid, ...snapshot.data() } })
+                }
+            })
+        unsubscribe.push(listener)
+    })
+}
+
+export function fetchUserChats() {
+    return ((dispatch) => {
+        let listener = firebase.firestore()
+            .collection("chats")
+            .where("users", "array-contains", firebase.auth().currentUser.uid)
+            .orderBy("lastMessageTimestamp", "desc")
+            .onSnapshot((snapshot) => {
+                let chats = snapshot.docs.map(doc => {
+                    const data = doc.data();
+                    const id = doc.id;
+                    return { id, ...data }
+                })
+
+                for (let i = 0; i < chats.length; i++) {
+                    let otherUserId;
+                    if (chats[i].users[0] == firebase.auth().currentUser.uid) {
+                        otherUserId = chats[i].users[1];
+                    } else {
+                        otherUserId = chats[i].users[0];
+                    }
+                    dispatch(fetchUsersData(otherUserId, false))
+                }
+
+                dispatch({ type: USER_CHATS_STATE_CHANGE, chats })
+            })
+        unsubscribe.push(listener)
+    })
+}
 export function fetchUserPosts() {
     return ((dispatch) => {
         firebase.firestore()
             .collection("posts")
             .doc(firebase.auth().currentUser.uid)
             .collection("userPosts")
-            .orderBy("creation", "asc")
+            .orderBy("creation", "desc")
             .get()
             .then((snapshot) => {
                 let posts = snapshot.docs.map(doc => {
@@ -45,9 +157,10 @@ export function fetchUserPosts() {
     })
 }
 
+
 export function fetchUserFollowing() {
     return ((dispatch) => {
-        firebase.firestore()
+        let listener = firebase.firestore()
             .collection("following")
             .doc(firebase.auth().currentUser.uid)
             .collection("userFollowing")
@@ -57,10 +170,11 @@ export function fetchUserFollowing() {
                     return id
                 })
                 dispatch({ type: USER_FOLLOWING_STATE_CHANGE, following });
-                for(let i = 0; i < following.length; i++){
+                for (let i = 0; i < following.length; i++) {
                     dispatch(fetchUsersData(following[i], true));
                 }
             })
+        unsubscribe.push(listener)
     })
 }
 
@@ -79,13 +193,10 @@ export function fetchUsersData(uid, getPosts) {
 
                         dispatch({ type: USERS_DATA_STATE_CHANGE, user });
                     }
-                    else {
-                        console.log('does not exist')
-                    }
                 })
-                if(getPosts){
-                    dispatch(fetchUsersFollowingPosts(uid));
-                }
+            if (getPosts) {
+                dispatch(fetchUsersFollowingPosts(uid));
+            }
         }
     })
 }
@@ -99,7 +210,7 @@ export function fetchUsersFollowingPosts(uid) {
             .orderBy("creation", "asc")
             .get()
             .then((snapshot) => {
-                const uid = snapshot.query._.C_.path.segments[1]
+                const uid = snapshot.docs[0].ref.path.split('/')[1];
                 const user = getState().usersState.users.find(el => el.uid === uid);
 
 
@@ -109,7 +220,7 @@ export function fetchUsersFollowingPosts(uid) {
                     return { id, ...data, user }
                 })
 
-                for(let i = 0; i< posts.length; i++){
+                for (let i = 0; i < posts.length; i++) {
                     dispatch(fetchUsersFollowingLikes(uid, posts[i].id))
                 }
                 dispatch({ type: USERS_POSTS_STATE_CHANGE, posts, uid })
@@ -120,7 +231,7 @@ export function fetchUsersFollowingPosts(uid) {
 
 export function fetchUsersFollowingLikes(uid, postId) {
     return ((dispatch, getState) => {
-        firebase.firestore()
+        let listener = firebase.firestore()
             .collection("posts")
             .doc(uid)
             .collection("userPosts")
@@ -128,14 +239,66 @@ export function fetchUsersFollowingLikes(uid, postId) {
             .collection("likes")
             .doc(firebase.auth().currentUser.uid)
             .onSnapshot((snapshot) => {
-                const postId = snapshot.ZE.path.segments[3];
+                const postId = snapshot.id;
 
                 let currentUserLike = false;
-                if(snapshot.exists){
+                if (snapshot.exists) {
                     currentUserLike = true;
                 }
 
                 dispatch({ type: USERS_LIKES_STATE_CHANGE, postId, currentUserLike })
             })
+        unsubscribe.push(listener)
     })
 }
+
+
+
+export function queryUsersByUsername(username) {
+    return ((dispatch, getState) => {
+        return new Promise((resolve, reject) => {
+            if (username.length == 0) {
+                resolve([])
+            }
+            firebase.firestore()
+                .collection('users')
+                .where('username', '>=', username)
+                .limit(10)
+                .get()
+                .then((snapshot) => {
+                    let users = snapshot.docs.map(doc => {
+                        const data = doc.data();
+                        const id = doc.id;
+                        return { id, ...data }
+                    });
+                    resolve(users);
+                })
+        })
+    })
+}
+
+
+export function deletePost(item) {
+    return ((dispatch, getState) => {
+        return new Promise((resolve, reject) => {
+            firebase.firestore()
+                .collection('posts')
+                .doc(firebase.auth().currentUser.uid)
+                .collection("userPosts")
+                .doc(item.id)
+                .delete()
+                .then(() => {
+                    resolve();
+                }).catch(() => {
+                    reject();
+                })
+        })
+    })
+}
+
+
+
+
+
+
+
